requirements.txt
# === Core Django ===
Django==5.0.9
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-filter==24.3

# === Database ===
psycopg[binary]==3.2.3
dj-database-url==2.2.0

# === Background tasks ===
celery==5.4.0
redis==5.1.1
django-celery-beat==2.7.0
django-celery-results==2.5.1

# === External services ===
httpx==0.27.2            # for ServiceM8, Gemini, Resend HTTP calls
google-generativeai==0.8.3
svix==1.34.0             # Resend webhook signature verification

# === Security / utils ===
cryptography==43.0.1     # encrypts SM8 tokens in settings table
python-decouple==3.8     # reads .env files
django-axes==7.0.0       # brute-force login protection

# === API docs ===
drf-spectacular==0.27.2

# === Dev tools ===
ipython==8.27.0
django-extensions==3.2.3
pytest==8.3.3
pytest-django==4.9.0
factory-boy==3.3.1
ruff==0.6.9